Add two boolean flag to prevent expand or collapse pass through anchored state.
In the current version, it is too easy to bypass the anchor level height (It will act like normal bottom sheet behavior).
If there have an intermediate page, it's not easy to scroll in anchor height and make the position locked.
Sometimes, It causes bad user experience.

This problem is caused by the method of shouldExpand and shouldCollapsed.
When the bottom sheet is below the anchor height, the 'shouldExpand' method will return
the value, which depends on the new top value(calculated by the current bottom sheet height and
user's scrolling velocity multiply the friction const value)
large then the anchor height, and make the bottom sheet expand,
so as the 'shouldCollapsed' method.

I add two boolean flags to prevent the behavior above.
Take the 'shouldExpand' for example, if the bottom sheet is above the anchor height but it's not in the anchor state.
It should scroll to anchor level first.
When it comes to collapsing situation, The bottom sheet should collapse from expanded state to anchor height then collapsed state.